Implement MFA type detection in GMAuth
Add logic to determine the type of multi-factor authentication
(MFA) based on the response data. Currently, the implementation
supports TOTP only, and an error is thrown if a different MFA
type is detected. This ensures that users are informed to
update their OnStar account settings if they are using a
non-supported MFA method.

Signed-off-by: Matt Nelson <metheos@gmail.com>